pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-web-assistant"
version = "1.0.0"
description = "Voice-controlled web browser assistant"
authors = [{name = "Tu Nombre", email = "tu@email.com"}]
license = {text = "MIT"}
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pyaudio>=0.2.11",
    "speechrecognition>=3.10.0",
    "gtts>=2.4.0",
    "pygame>=2.5.0",
    "openai>=1.3.0",
    "pillow>=10.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
]
